+++ academy-backend/src/main/java/uz/jurayev/academy/repository/AttachmentRepository.java
@@ -3,5 +3,5 @@ package uz.jurayev.academy.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import uz.jurayev.academy.domain.Attachment;
 
-public interface AttachmentRepository extends JpaRepository<Attachment, Long> {
+public interface AttachmentRepository extends JpaRepository<Attachment, Integer> {
 }
