+++ academy-backend/src/main/java/uz/jurayev/academy/repository/CreativePotentialRepository.java
@@ -3,5 +3,5 @@ package uz.jurayev.academy.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import uz.jurayev.academy.domain.CreativePotential;
 
-public interface CreativePotentialRepository extends JpaRepository<CreativePotential, Long> {
+public interface CreativePotentialRepository extends JpaRepository<CreativePotential, Integer> {
 }
