+++ academy-backend/src/main/java/uz/jurayev/academy/repository/NationalityRepository.java
@@ -3,5 +3,5 @@ package uz.jurayev.academy.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import uz.jurayev.academy.domain.Nationality;
 
-public interface NationalityRepository extends JpaRepository<Nationality, Long> {
+public interface NationalityRepository extends JpaRepository<Nationality, Integer> {
 }
