+++ academy-backend/src/main/java/uz/jurayev/academy/repository/TutorRepository.java
@@ -3,5 +3,5 @@ package uz.jurayev.academy.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import uz.jurayev.academy.domain.Tutor;
 
-public interface TutorRepository extends JpaRepository<Tutor, Long> {
+public interface TutorRepository extends JpaRepository<Tutor, Integer> {
 }
