+++ academy-backend/src/main/java/uz/jurayev/academy/repository/UniversityRepository.java
@@ -3,6 +3,6 @@ package uz.jurayev.academy.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import uz.jurayev.academy.domain.University;
 
-public interface UniversityRepository extends JpaRepository<University, Long> {
+public interface UniversityRepository extends JpaRepository<University, Integer> {
     boolean existsByName(String name);
 }
