+++ academy-backend/src/main/java/uz/jurayev/academy/service/CountryService.java
@@ -1,6 +1,5 @@
 package uz.jurayev.academy.service;
 
-
 import uz.jurayev.academy.domain.Country;
 import uz.jurayev.academy.model.Result;
 import java.util.List;
@@ -9,8 +8,8 @@ public interface CountryService {
 
     Result addCountry(Country country);
     List<Country> getAll(int page, int size);
-    Country getOne(Long id);
-    Country edit(Long id, Country country);
-    boolean delete(Long id);
+    Country getOne(Integer id);
+    Country edit(Integer id, Country country);
+    boolean delete(Integer id);
 
 }
